refactor: rename OutlierAnalysisTable directory
